+++ src/components/TodoAddInput.jsx
@@ -1,20 +1,7 @@
-import React, { useState } from 'react';
+import React from 'react';
 
-export default function TodoAddInput() {
-  const [state, setState] = useState({
-    inputValue: '',
-  });
-
-  const { inputValue } = state;
-
-  function handleChange({ target }) {
-    const { value } = target;
-    setState({
-      inputValue: value,
-    });
-  }
-
+export default function TodoAddInput({ inputValue, onChange }) {
   return (
-    <input id="todo-input" value={inputValue} onChange={(e) => handleChange(e)} />
+    <input value={inputValue} onChange={(e) => onChange(e)} />
   );
 }
